# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Immobilière",
        "description": "API pour la gestion des biens immobiliers",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/visit/store": {
            "post": {
                "tags": [
                    "Visites"
                ],
                "summary": "Créer une nouvelle demande de visite",
                "operationId": "7de35ba7b1c85bb30b3c3350f67ac141",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bien_id",
                                    "nom",
                                    "email",
                                    "telephone",
                                    "date_visite",
                                    "heure_visite"
                                ],
                                "properties": {
                                    "bien_id": {
                                        "description": "ID du bien à visiter",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nom": {
                                        "description": "Nom du visiteur",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email du visiteur",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john@example.com"
                                    },
                                    "telephone": {
                                        "description": "Numéro de téléphone ivoirien (mobile)",
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "07 98 27 89 81"
                                    },
                                    "date_visite": {
                                        "description": "Date souhaitée pour la visite (au format YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "heure_visite": {
                                        "description": "Heure souhaitée pour la visite",
                                        "type": "string",
                                        "example": "14:00"
                                    },
                                    "message": {
                                        "description": "Message supplémentaire",
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Je souhaite visiter en matinée si possible",
                                        "nullable": true
                                    },
                                    "statut": {
                                        "description": "Statut de la visite (par défaut: en attente)",
                                        "type": "string",
                                        "enum": [
                                            "en attente",
                                            "confirmée",
                                            "effectuée",
                                            "annulée"
                                        ],
                                        "example": "en attente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Demande de visite créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Demande de visite enregistrée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "visite": {
                                                    "$ref": "#/components/schemas/Visite"
                                                },
                                                "bien": {
                                                    "$ref": "#/components/schemas/Bien"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "L'email doit être une adresse valide"
                                                ],
                                                "telephone": [
                                                    "Le numéro doit être un mobile ivoirien valide"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création de la visite"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Message d'erreur détaillé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Bien": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titre": {
                        "type": "string",
                        "example": "Appartement spacieux"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "Appartement"
                    },
                    "prix": {
                        "type": "number",
                        "format": "float",
                        "example": 750.5
                    },
                    "surface": {
                        "type": "integer",
                        "example": 80
                    },
                    "commune": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "status": {
                        "type": "string",
                        "example": "Disponible"
                    }
                },
                "type": "object"
            },
...
            "Visite": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "date_visite": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-11-15"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}